examples/requirements.txt
httpx>=0.26.0
python-dotenv>=1.0.0
rich>=13.0.0
pillow>=10.0.0
langchain-anthropic==1.4.0
langchain-mcp-adapters==0.2.2
langgraph==1.1.4
